added generic service
